% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{classify}
\alias{classify}
\title{Classify image(s) (todo: add progress meter)}
\usage{
classify(images, ...)
}
\arguments{
\item{images}{image paths (either url or server path)}

\item{...}{arguments passed on to classify_one(...)}
}
\description{
...
}

