% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{start_api}
\alias{start_api}
\title{Start File Server API}
\usage{
start_api(server_path = "/media/reinholdsson/data/digits",
  host = "localhost", port = 5000)
}
\arguments{
\item{host}{digits server host, e.g. localhost}

\item{port}{digits server port, e.g. 5000}
}
\description{
...
}

